Task 236: consolidate services into single package
